+++ config/default.js
@@ -323,8 +323,8 @@ defaultConfig.security.sessions.cookieName = process.env.API_SECURITY_SESSIONS_C
 defaultConfig.security.sessions.onlySecure = process.env.API_SECURITY_SESSIONS_ONLYSECURE || defaultConfig.security.sessions.onlySecure;
 defaultConfig.authorization['jwt-secret'] = process.env.API_AUTHORIZATION_JWTSECRET || defaultConfig.authorization['jwt-secret'];
 defaultConfig.authorization['auth-server-url'] = process.env.AUTH_API_URL || defaultConfig.authorization['auth-server-url'];
-defaultConfig.authorization["auth-client-id"] = process.env.USER_API_CLIENT_ID || defaultConfig.authorization["auth-client-id"];
-defaultConfig.authorization["auth-client-secret"] = process.env.USER_API_CLIENT_SECRET || defaultConfig.authorization["auth-client-secret"];
+defaultConfig.authorization["auth-client-id"] = process.env.AUTH_FIRST_CLIENT_ID || defaultConfig.authorization["auth-client-id"];
+defaultConfig.authorization["auth-client-secret"] = process.env.AUTH_FIRST_CLIENT_SECRET || defaultConfig.authorization["auth-client-secret"];
 
 try {
   defaultConfig.authorization['fixed-auth-tokens'] = JSON.parse(process.env.API_AUTHORIZATION_FIXEDAUTHTOKENS) || defaultConfig.authorization['fixed-auth-tokens'];