+++ src/db.js
@@ -7,7 +7,7 @@ var dbConfig  = config.get('database');
 
 // newer versions of mysql (8+) have changed GeomFromText to ST_GeomFromText
 // this is a fix for sequalize
-if (dbConfig.mysqlSTGeoMode) {
+if (dbConfig.mysqlSTGeoMode || process.env.MYSQL_ST_GEO_MODE === 'on') {
 	const wkx = require('wkx')
 	Sequelize.GEOMETRY.prototype._stringify = function _stringify(value, options) {
 	  return `ST_GeomFromText(${options.escape(wkx.Geometry.parseGeoJSON(value).toWkt())})`;
@@ -35,7 +35,7 @@ var sequelize = new Sequelize(dbConfig.database, dbConfig.user, dbConfig.passwor
 	},
 	timeZone       : config.timeZone,
 	logging        : require('debug')('app:db:query'),
- 	//logging				 : console.log,
+ 	// logging				 : console.log,
 	typeValidation : true,
 
 	define: {