+++ src/middleware/site.js
@@ -1,5 +1,6 @@
 const db = require('../db');
 const createError = require('http-errors');
+const siteConfig = require('../lib/siteConfig');
 
 const getSiteId = (path) => {
   const match = path.match(/\/site\/(\d+)?\//);
@@ -25,11 +26,17 @@ module.exports = function( req, res, next ) {
   return db.Site
   	.findOne({ where })
   	.then(function( found ) {
-  		if (!found) return next(new createError('400', 'Site niet gevonden'));
+      if (!found) {
+        console.log('Site not found for siteId query: ', where);
+        return next(new createError('400', 'Site niet gevonden for siteId: '+ siteId));
+      }
+      siteConfig.setFromSite(found);
   		req.site = found;
   		next();
+      return null;
   	})
   	.catch( err => {
   		next(err)
+      return null;
   	});
 }