+++ src/models/Site.js
@@ -366,6 +366,22 @@ module.exports = function( db, sequelize, DataTypes ) {
 
 				}
 			},
+      users: {
+				type: 'object',
+				subset: {
+          extraDataMustBeDefined: {
+					  type: 'boolean',
+            default: false,
+				  },
+          extraData: {
+					  type: 'object',
+				  },
+          canCreateNewUsers: {
+					  type: 'boolean',
+					  default: true,
+				  },
+				},
+      },
 			votes: {
 				type: 'object',
 				subset: {
@@ -408,7 +424,7 @@ module.exports = function( db, sequelize, DataTypes ) {
 
 					voteType: {
 						type: 'enum',
-						values: ['likes', 'count', 'budgeting', 'budgeting-per-theme'],
+						values: ['likes', 'count', 'budgeting', 'count-per-theme', 'budgeting-per-theme'],
 						default: 'likes',
 					},
 
@@ -428,12 +444,12 @@ module.exports = function( db, sequelize, DataTypes ) {
 
 					maxIdeas: {
 						type: 'int',
-						default: undefined,
+						default: 100,
 					},
 
 					minIdeas: {
 						type: 'int',
-						default: undefined,
+						default: 1,
 					},
 
 					minBudget: {