+++ src/models/User.js
@@ -6,6 +6,7 @@ const Password      = require('../lib/password');
 const sanitize      = require('../util/sanitize');
 const userHasRole 	= require('../lib/sequelize-authorization/lib/hasRole');
 const getExtraDataConfig = require('../lib/sequelize-authorization/lib/getExtraDataConfig');
+const roles = require('../lib/sequelize-authorization/lib/roles');
 
 // For detecting throwaway accounts in the email address validation.
 var emailBlackList = require('../../config/mail_blacklist')
@@ -168,6 +169,26 @@ module.exports = function( db, sequelize, DataTypes ) {
 			set          : function( value ) {
 				this.setDataValue('lastName', sanitize.noTags(value));
 			}
+		},
+
+		listableByRole: {
+			type: DataTypes.ENUM('admin', 'moderator', 'editor', 'member', 'anonymous', 'all'),
+			defaultValue: null,
+			auth:  {
+        viewableBy: ['editor','owner'],
+				updateableBy: ['editor', 'owner'],
+			},
+			allowNull: true,
+		},
+
+		detailsViewableByRole: {
+			type: DataTypes.ENUM('admin', 'moderator', 'editor', 'member', 'anonymous', 'all'),
+			defaultValue: null,
+			auth:  {
+        viewableBy: ['editor','owner'],
+				updateableBy: ['editor', 'owner'],
+			},
+			allowNull: true,
 		},
 
 		phoneNumber: {
@@ -342,11 +363,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 			unique: true
 		}],*/
 
-		includeSite: {
-			include: [{
-				model: db.Site,
-			}]
-		},
 
 		validate: {
 			hasValidUserRole: function() {
@@ -383,12 +399,81 @@ module.exports = function( db, sequelize, DataTypes ) {
 	});
 
 	User.scopes = function scopes() {
+
 		return {
 			includeSite: {
         include: [{
           model: db.Site,
         }]
       },
+
+			onlyListable: function (userId, userRole = 'all') {
+
+        // todo: hij kan alleen tegen een enkelvoudige listableBy
+        // todo: owner wordt nu altijd toegevoegd, dat moet alleen als die in listableBy staat, maar zie vorige regel
+        // todo: gelijkttrekken met Idea.onlyVisible: die is nu exclusive en deze inclusive
+
+        let requiredRole = this.auth && this.auth.listableBy || 'all';
+				
+        // if requiredRole == all then listableByRole is not relevant and neither is userRole
+        if (requiredRole === 'all') return;
+
+        // if requiredRole != all then listableByRole is allowing
+
+        // null should be seen as requiredRole
+        let requiredRoleEscaped = sequelize.escape(requiredRole);
+        let rolesEscaped = sequelize.escape(roles[userRole])
+        let nullCondition = `${requiredRoleEscaped} IN (${rolesEscaped})`;
+
+        let where;
+				if (userId) {
+          where = sequelize.or(
+            { id: userId }, // owner
+            { listableByRole: roles[userRole] || 'none' }, // allow when userRole is good enough
+            sequelize.and( // or null and userRole is at least requiredRole
+              { listableByRole: null },
+              sequelize.literal(nullCondition)
+            ),
+          )
+        } else {
+          where = sequelize.or(
+            { listableByRole: roles[userRole] || 'none' }, // allow when userRole is good enough
+            sequelize.and( // or null and userRole is at least requiredRole
+              { listableByRole: null },
+              sequelize.literal(nullCondition)
+            ),
+          )
+        }
+
+        return { where };
+
+			},
+
+			includeVote: {
+				include: [{
+					model: db.Vote,
+				}]
+			},
+
+
+			onlyVisible: function (userId, userRole) {
+				if (userId) {
+					return {
+						where: sequelize.or(
+							{ id: userId },
+							{ viewableByRole: 'all' },
+							{ viewableByRole: roles[userRole] || 'all' },
+						)
+					};
+				} else {
+					return {
+						where: sequelize.or(
+							{ viewableByRole: 'all' },
+							{ viewableByRole: roles[userRole]  || 'all' },
+						)
+					};
+				}
+ 			},
 
 		}
 	}
@@ -399,86 +484,6 @@ module.exports = function( db, sequelize, DataTypes ) {
 		this.hasMany(models.Vote);
 		this.hasMany(models.Argument);
 		this.belongsTo(models.Site);
-	}
-
-	User.findByCredentials = function( email, password ) {
-		if( !email || !password ) {
-			return Promise.reject(createError(400, 'Incomplete credentials'));
-		}
-
-		return this.findOne({where: {email: email}}).then(function( user ) {
-			if( !user || !user.authenticate(password) ) {
-				// TODO: AuthenticationError
-				throw createError(403, 'Login failed');
-			} else {
-				return user;
-			}
-		});
-	}
-
-	User.findMember = function( email ) {
-		if( !email ) {
-			return Promise.reject(createError(400, 'Geen emailadres ingevuld'));
-		}
-
-		return this.findOne({where: {
-			email : email
-		}});
-	}
-
-	User.registerAnonymous = function(values) {
-		values.role = 'anonymous';
-		return this.create(values);
-	}
-
-	// `User.currentUser` is the user instance that the system loaded for
-	// this session.
-	User.registerMember = function( currentUser, email ) {
-		var data = {
-			role  : 'member',
-			email : email
-		};
-
-		return User.findMember(email)
-			.then(function( existingUser ) {
-				if( existingUser ) {
-					// If there is already a member with this email address,
-					// just return that one. This is probably someone who
-					// actually wants to login.
-					return existingUser;
-				} else if( currentUser.isAnonymous() ) {
-					// If the current user is anonymous, it means they already
-					// created an 'account' by filling in their zipcode during
-					// voting. Upgrade this account to a member.
-					return currentUser.update(data);
-				} else {
-					// No existing user, and the current user is of type unknown.
-					// Members should not be able to reach this point; the ACL
-					// prevents it.
-					return User.create(data);
-				}
-			});
-	}
-
-	User.prototype.completeRegistration = function( data ) {
-		var self = this;
-		var filtered = pick(data, [
-			'firstName', 'lastName', 'zipCode', 'postcode', 'gender', 'password', 'signedUpForNewsletter'
-		]);
-		filtered.complete = true;
-		if (self.role === 'anonymous') {
-			filtered.complete = 'member';
-		}
-		return this.update(filtered)
-			.catch(function( error ) {
-				// We need to set `complete` initially for the `isValidMember`
-				// validation to work correctly. There was an error completing
-				// registration however, so we unset `complete` again. The
-				// other properties can remain in memory so they're refilled into
-				// the form together with the error message(s).
-				self.set('complete', false);
-				throw error;
-			});
 	}
 
 	User.prototype.authenticate = function( password ) {
@@ -551,6 +556,14 @@ module.exports = function( db, sequelize, DataTypes ) {
     createableBy: 'editor',
     updateableBy: ['editor','owner'],
     deleteableBy: ['editor','owner'],
+
+		/*canView: function(user, self) {
+			if (self && self.viewableByRole && self.viewableByRole != 'all' ) {
+				return userHasRole(user, [ self.viewableByRole, 'owner' ], self.userId)
+			} else {
+				return true
+			}
+		},*/
   }
 
 	return User;